+++ packages/platform-server/src/location.ts
@@ -13,29 +13,29 @@ import {
   PlatformLocation,
   ɵgetDOM as getDOM,
 } from '@angular/common';
-import {inject, Inject, Injectable, Optional, ɵWritable as Writable} from '@angular/core';
+import {inject, Injectable, ɵWritable as Writable} from '@angular/core';
 import {Subject} from 'rxjs';
 
-import {INITIAL_CONFIG, PlatformConfig} from './tokens';
+import {INITIAL_CONFIG} from './tokens';
 
-const RESOLVE_PROTOCOL = 'resolve:';
-
-function parseUrl(urlStr: string): {
+function parseUrl(
+  urlStr: string,
+  origin: string,
+): {
   hostname: string;
   protocol: string;
   port: string;
   pathname: string;
   search: string;
   hash: string;
+  href: string;
 } {
-  const {hostname, protocol, port, pathname, search, hash} = new URL(
-    urlStr,
-    RESOLVE_PROTOCOL + '//',
-  );
+  const {hostname, protocol, port, pathname, search, hash, href} = new URL(urlStr, origin);
 
   return {
     hostname,
-    protocol: protocol === RESOLVE_PROTOCOL ? '' : protocol,
+    href,
+    protocol,
     port,
     pathname,
     search,
@@ -65,14 +65,14 @@ export class ServerPlatformLocation implements PlatformLocation {
       return;
     }
     if (config.url) {
-      const url = parseUrl(config.url);
+      const url = parseUrl(config.url, this._doc.location.origin);
       this.protocol = url.protocol;
       this.hostname = url.hostname;
       this.port = url.port;
       this.pathname = url.pathname;
       this.search = url.search;
       this.hash = url.hash;
-      this.href = this._doc.location.href;
+      this.href = url.href;
     }
   }
 
@@ -114,9 +114,11 @@ export class ServerPlatformLocation implements PlatformLocation {
 
   replaceState(state: any, title: string, newUrl: string): void {
     const oldUrl = this.url;
-    const parsedUrl = parseUrl(newUrl);
+    const parsedUrl = parseUrl(newUrl, this._doc.location.origin);
     (this as Writable<this>).pathname = parsedUrl.pathname;
     (this as Writable<this>).search = parsedUrl.search;
+    (this as Writable<this>).href = parsedUrl.href;
+    (this as Writable<this>).protocol = parsedUrl.protocol;
     this.setHash(parsedUrl.hash, oldUrl);
   }
