+++ packages/platform-server/test/integration_spec.ts
@@ -752,108 +752,6 @@ class HiddenModule {}
       expect(img.attributes['src'].value).toEqual('link');
     });
 
-    describe('PlatformLocation', () => {
-      it('is injectable', async () => {
-        const platform = platformServer([
-          {provide: INITIAL_CONFIG, useValue: {document: '<app></app>'}},
-        ]);
-        const appRef = await platform.bootstrapModule(ExampleModule);
-        const location = appRef.injector.get(PlatformLocation);
-        expect(location.pathname).toBe('/');
-        platform.destroy();
-      });
-      it('is configurable via INITIAL_CONFIG', async () => {
-        const platform = platformServer([
-          {
-            provide: INITIAL_CONFIG,
-            useValue: {
-              document: '<app></app>',
-              url: 'http://test.com/deep/path?query#hash',
-            },
-          },
-        ]);
-
-        const appRef = await platform.bootstrapModule(ExampleModule);
-
-        const location = appRef.injector.get(PlatformLocation);
-        expect(location.pathname).toBe('/deep/path');
-        expect(location.search).toBe('?query');
-        expect(location.hash).toBe('#hash');
-      });
-
-      it('parses component pieces of a URL', async () => {
-        const platform = platformServer([
-          {
-            provide: INITIAL_CONFIG,
-            useValue: {
-              document: '<app></app>',
-              url: 'http://test.com:80/deep/path?query#hash',
-            },
-          },
-        ]);
-
-        const appRef = await platform.bootstrapModule(ExampleModule);
-
-        const location = appRef.injector.get(PlatformLocation);
-        expect(location.hostname).toBe('test.com');
-        expect(location.protocol).toBe('http:');
-        expect(location.port).toBe('');
-        expect(location.pathname).toBe('/deep/path');
-        expect(location.search).toBe('?query');
-        expect(location.hash).toBe('#hash');
-      });
-
-      it('handles empty search and hash portions of the url', async () => {
-        const platform = platformServer([
-          {
-            provide: INITIAL_CONFIG,
-            useValue: {
-              document: '<app></app>',
-              url: 'http://test.com/deep/path',
-            },
-          },
-        ]);
-
-        const appRef = await platform.bootstrapModule(ExampleModule);
-
-        const location = appRef.injector.get(PlatformLocation);
-        expect(location.pathname).toBe('/deep/path');
-        expect(location.search).toBe('');
-        expect(location.hash).toBe('');
-      });
-
-      it('pushState causes the URL to update', async () => {
-        const platform = platformServer([
-          {provide: INITIAL_CONFIG, useValue: {document: '<app></app>'}},
-        ]);
-
-        const appRef = await platform.bootstrapModule(ExampleModule);
-        const location = appRef.injector.get(PlatformLocation);
-        location.pushState(null, 'Test', '/foo#bar');
-        expect(location.pathname).toBe('/foo');
-        expect(location.hash).toBe('#bar');
-        platform.destroy();
-      });
-
-      it('allows subscription to the hash state', (done) => {
-        const platform = platformServer([
-          {provide: INITIAL_CONFIG, useValue: {document: '<app></app>'}},
-        ]);
-        platform.bootstrapModule(ExampleModule).then((appRef) => {
-          const location: PlatformLocation = appRef.injector.get(PlatformLocation);
-          expect(location.pathname).toBe('/');
-          location.onHashChange((e: any) => {
-            expect(e.type).toBe('hashchange');
-            expect(e.oldUrl).toBe('/');
-            expect(e.newUrl).toBe('/foo#bar');
-            platform.destroy();
-            done();
-          });
-          location.pushState(null, 'Test', '/foo#bar');
-        });
-      });
-    });
-
     describe('render', () => {
       let doc: string;
       let expectedOutput =
